Extract working_memory_messages.rs from turn.rs (parent #680)
Sixteenth vertical-slice extraction: move the working-memory + repo-
context prompt message builders out of turn.rs into a new
src/agent/loop_run/working_memory_messages.rs sibling module.

Migrated (4 pub(super) free fns):
- refresh_working_memory(&mut Agent) — re-extract plan constraints
  from the active plan file and replace the session's working-memory
  constraint list.
- working_memory_message(&mut Agent) -> Option<ConversationMessage> —
  render the working-memory snippet (with per-prompt precaution
  filtering) as a ConversationMessage when mode policy permits.
- answer_only_fallback_response(&Agent) -> String — AnswerOnly mode's
  canned read-only response, selected from the current request text.
- repo_context_message(&mut Agent) -> Option<ConversationMessage> —
  produce the cached repo-context system message (RepoGraph +
  suspected files + change history), bypassing when the cache key
  matches.

No facade re-export (DR3-001). Caller updates:
- build_request_messages.rs (×2): agent.working_memory_message() /
  agent.repo_context_message() → free-fn form.
- actor_loop_flow.rs (×2): agent.answer_only_fallback_response() →
  super::working_memory_messages::answer_only_fallback_response(agent).

`cargo fix --tests` pruned 3 newly-unused imports from turn.rs's prelude.

turn.rs: 3,704 -> 3,571 (-133 LOC)
Cumulative from 39,489: -35,918 LOC (-90.9%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
